fix: address code-review findings for viv loader path
Fix correctness, concurrency, and structural issues found in the
feat/viv-loader review:

- Infer OF (not OL) for PointCoordinatesData when DICOM JSON omits VR;
  harden Range streaming against ignored Range headers / 416 EOF; stop
  silently completing truncated downloads; remove POLYLINE off-by-one
  phantom vertices; keep finite zero coordinates as valid.
- Deduplicate prefix emission into bulkPrefixEmitter and coordinate
  transform/validity helpers into annotationCoords (shared with the
  center-out worker kernel). Transfer only first-vertex XY to the worker.
- Make DicomLoader ICC sync single-flight and dispose-guarded; promise-
  memoize viewer creation; LRU + generation-guard the tile cache; skip
  out-of-range synthetic dyadic tiles; suppress only bridged tile aborts.
- Extract bulk-annotation orchestration into useVivBulkAnnotations with
  per-group hydrate generations, error-phase reporting, constant scatter
  accessors, and binary streaming centroid data; shrink VivSlideViewport.
- Apply persisted ICC preference on classic slide switch; reset viv
  annotation panel state on series change; restore sidebar contrast under
  disabled antd menu items; purify preference reads and share one storage
  listener; fix craco @vivjs include and TypedArray typings; tune the
  XHR retry hook (idempotent methods only, shorter backoff).
